requirements.txt
Django~=4.1.0
psycopg2~=2.9.0
djangorestframework~=3.14.0
drf-spectacular~=0.25.0
django-widget-tweaks~=1.4.0
